cognitivehealthmonitoring/data_push_scripts/requirements.txt
azure-cosmos>=4.0.0
